Unit Testing and Local Variables
Added unit tests to verify the correctness of the various transactional
operations. Added in the concept of local variables (LOAD and STORE
operators) which are mutable within the execution of a transaction but
are not visible to other transactions. This enabled loops and other nice
properties. Still need to work on scoping local variables.