Updated caching
Mitigates the cache thrashing problem, by only invalidating the specific
keys that cause a conflict and not all keys that participate in the
transaction. There is still the vulnerability for thrashing, but at
least now it will only thrash on the keys that are actually under
contention. For very high contention situations, it might better to do
away with the cache entirely.