Caustic Service
Implemented a convenient interface for standing up and connecting to
instances and clusters of instances. Also includes a codec to more
easily construct Caustic transaction (the Thrift-generated interface is
hideous, because Java does not have native support for unions) and an
implementation of service discovery to allow clients to dynamically
execute requests on clusters of servers. Closes #32.