Fixed Automatic Unregistration
Right now, server instances are only removed when the ZooKeeper session
expires because they rely on guarantees about ephemeral sequential
nodes. Furthermore, it is possible that server instances can restart
before ZooKeeper has a chance to remove the expired ephemeral nodes as
found
https://blog.box.com/blog/a-gotcha-when-using-zookeeper-ephemeral-nodes/.

To prevent this, we add an "unregister" function which instances can
manually call when they terminate and which the registry automatically
calls before registering an instance.