Improved Cache Eviction
LocalCache is currently bounded by the number of entries and not by its
total memory utilization. Because cache entries are not homogenous in
length, the total size of the cache is effectively unbounded. Ran some
experiments that closely approximate the size of cache entries and
changed the LocalCache to be bounded by total memory utilization
instead.