Removed Common Package
Much of the need for the common package has been mitigated by the
caustic-service package and the creation of the top-level sandbox
repository as a dumping ground for random projects/ideas. Removing the
common package will reduce the number of artifacts published to Maven
Central and removes unnecessary dependencies on various 3rdparty
libraries.