Documentation Overhaul and SQLDatabase Refactor
Added more consistent documentation to the runtime package. Fixed test
failures with the MySQLDialect and PostgreSQLDialect. Added comments
to explain why queries were written that way. Improved the performance
of upsert by performing the upserts in bulk instead of one at a time.
Fixed the test harness for SQLDatabases to make it easier to add
additional databases in the future.