Branch Prefetching
Made it so dependencies are only taken on keys in the branch that is
taken. While both branches are prefetched, a dependency is only required
when the data is actually read or written. This should reduce
significantly reduce the size of transactions, which has consequences
for performance because we know that the runtime scales with transaction
size. Closes #60.